C program that prints the size of various types

+++ 0x00-hello_world/6-size.c
@@ -12,7 +12,7 @@ int main(void)
 
 	long longVar;
 
-	long longLongVar;
+	long long longLongVar;
 
 	float floatVar;
